Clear the clang-tidy findings on main
Seven findings at the pinned clang-tidy 20.1.0, none of them a bug in
behaviour, all of them things the gate is right to ask about.

to_native in signal_posix.c was a switch of eleven cases that each
returned a different constant. The clone detector compares the shape of
a branch and not its constants, so it read them as eleven copies. It is
now a table of pairs and a loop, which is also how a mapping reads best.

The static assertion in errno.c cast an int difference to size_t, which
the widening cast check reads as a cast in the wrong place. The count is
cast down to int instead, which is the side that is known to be small.

The signal stack floor multiplied two ints into an int64_t. It is still
64 kilobytes, spelled with INT64_C so the multiply happens in the wide
type.

type_by_name was declared with a parameter called name and defined with
one called q. The definition now uses name, which is what the header
documents, and the bare type name split out of it is called base.

The FNV constants in registry.c get uppercase suffixes and waitgroup.c
loses a cast to the type the expression already had.

Last, the ELF linker names the two ends of the type section
__start_burrowtype and __stop_burrowtype, which are reserved spellings
that are not ours to change. They go on the allowlist in .clang-tidy with
the feature macros already there, rather than behind a NOLINT.

+++ src/pal/errno.c
@@ -79,8 +79,8 @@ static const char *const errno_names[] = {
     "unknown error",
 };
 
-_Static_assert(sizeof errno_names / sizeof errno_names[0] ==
-                   (size_t)(PAL_EOTHER - PAL_EPERM + 1),
+_Static_assert((int)(sizeof errno_names / sizeof errno_names[0]) ==
+                   PAL_EOTHER - PAL_EPERM + 1,
                "errno_names and the PalErrno enum have drifted apart");
 
 const char *pal_errno_string(PalErrno e) {
